devtools/client/shared/components/reps/test/mochitest/test_reps_object-with-text.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test ObjectWithText rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - ObjectWithText</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    const { REPS } = browserRequire("devtools/client/shared/components/reps/load-reps");
    let { Rep, ObjectWithText } = REPS;

    let gripStub = {
      "type": "object",
      "class": "CSSStyleRule",
      "actor": "server1.conn3.obj273",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 0,
      "preview": {
        "kind": "ObjectWithText",
        "text": ".Shadow"
      }
    };

    // Test that correct rep is chosen
    const renderedRep = shallowRenderComponent(Rep, { object: gripStub });
    is(renderedRep.type, ObjectWithText.rep, `Rep correctly selects ${ObjectWithText.rep.displayName}`);

    // Test rendering
    const renderedComponent = renderComponent(ObjectWithText.rep, { object: gripStub });
    is(renderedComponent.textContent, "\".Shadow\"", "ObjectWithText rep has expected text content");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
